introduce-weight-updater-controller(absorb-move): Hand the remaining weight ops over to WeightUpdaterController
Pure physical move per MECH_COMMIT_SPLIT. Cut the @staticmethod weight methods
(``init_weights_update_group`` / ``destroy_weights_update_group`` /
``update_weights_from_distributed`` / ``update_weights_from_tensor`` /
``update_weights_from_ipc`` / ``get_weights_by_name`` /
``release_memory_occupation`` / ``resume_memory_occupation`` /
``check_weights``) from TokenizerControlMixin and paste into
WeightUpdaterController (drop @staticmethod, replace
``self: "WeightUpdaterController"`` -> plain ``self``, fold the class-qualified
version-bump call back to the sibling ``self._update_weight_version_if_provided``).
Add the imports the moved bodies need. Caller prefix replacement:
``TokenizerManager.<method>(<facade>.weight_updater_controller, ...)`` ->
``<facade>.weight_updater_controller.<method>(...)`` in engine.py and
http_server.py.

Refactor chain ID: wuc-absorb-move
PR-Title: Extract WeightUpdaterController from TokenizerManager
