+++ python/sglang/srt/entrypoints/http_server.py
@@ -1168,8 +1168,10 @@ async def init_weights_update_group(
     obj: InitWeightsUpdateGroupReqInput, request: Request
 ):
     """Initialize the parameter update group."""
-    success, message = await TokenizerManager.init_weights_update_group(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message = (
+        await _global_state.tokenizer_manager.weight_updater_controller.init_weights_update_group(
+            obj, request
+        )
     )
     content = {"success": success, "message": message}
     if success:
@@ -1184,8 +1186,10 @@ async def destroy_weights_update_group(
     obj: DestroyWeightsUpdateGroupReqInput, request: Request
 ):
     """Destroy the parameter update group."""
-    success, message = await TokenizerManager.destroy_weights_update_group(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message = (
+        await _global_state.tokenizer_manager.weight_updater_controller.destroy_weights_update_group(
+            obj, request
+        )
     )
     content = {"success": success, "message": message}
     return ORJSONResponse(
@@ -1205,8 +1209,10 @@ async def update_weights_from_tensor(
     3. Any binary data in the named tensors should be base64 encoded.
     """
 
-    success, message = await TokenizerManager.update_weights_from_tensor(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message = (
+        await _global_state.tokenizer_manager.weight_updater_controller.update_weights_from_tensor(
+            obj, request
+        )
     )
 
     content = {"success": success, "message": message}
@@ -1221,8 +1227,10 @@ async def update_weights_from_distributed(
     obj: UpdateWeightsFromDistributedReqInput, request: Request
 ):
     """Update model parameter from distributed online."""
-    success, message = await TokenizerManager.update_weights_from_distributed(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message = (
+        await _global_state.tokenizer_manager.weight_updater_controller.update_weights_from_distributed(
+            obj, request
+        )
     )
 
     content = {"success": success, "message": message}
@@ -1236,8 +1244,10 @@ async def update_weights_from_distributed(
 @auth_level(AuthLevel.ADMIN_OPTIONAL)
 async def update_weights_from_ipc(obj: UpdateWeightsFromIPCReqInput, request: Request):
     """Update the weights from IPC (Inter-Process Communication) for checkpoint-engine integration."""
-    success, message = await TokenizerManager.update_weights_from_ipc(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message = (
+        await _global_state.tokenizer_manager.weight_updater_controller.update_weights_from_ipc(
+            obj, request
+        )
     )
 
     content = {"success": success, "message": message}
@@ -1290,8 +1300,8 @@ async def update_weight_version(obj: UpdateWeightVersionReqInput, request: Reque
 async def get_weights_by_name(obj: GetWeightsByNameReqInput, request: Request):
     """Get model parameter by name."""
     try:
-        ret = await TokenizerManager.get_weights_by_name(
-            _global_state.tokenizer_manager.weight_updater_controller, obj, request
+        ret = await _global_state.tokenizer_manager.weight_updater_controller.get_weights_by_name(
+            obj, request
         )
         if ret is None:
             return _create_error_response("Get parameter by name failed")
@@ -1308,8 +1318,8 @@ async def release_memory_occupation(
 ):
     """Release GPU memory occupation temporarily."""
     try:
-        await TokenizerManager.release_memory_occupation(
-            _global_state.tokenizer_manager.weight_updater_controller, obj, request
+        await _global_state.tokenizer_manager.weight_updater_controller.release_memory_occupation(
+            obj, request
         )
     except Exception as e:
         return _create_error_response(e)
@@ -1322,8 +1332,8 @@ async def resume_memory_occupation(
 ):
     """Resume GPU memory occupation."""
     try:
-        await TokenizerManager.resume_memory_occupation(
-            _global_state.tokenizer_manager.weight_updater_controller, obj, request
+        await _global_state.tokenizer_manager.weight_updater_controller.resume_memory_occupation(
+            obj, request
         )
     except Exception as e:
         return _create_error_response(e)
@@ -1336,8 +1346,10 @@ async def check_weights(
 ):
     if obj is None:
         obj = CheckWeightsReqInput()
-    success, message, ranks, per_engine_checksum = await TokenizerManager.check_weights(
-        _global_state.tokenizer_manager.weight_updater_controller, obj, request
+    success, message, ranks, per_engine_checksum = (
+        await _global_state.tokenizer_manager.weight_updater_controller.check_weights(
+            obj, request
+        )
     )
     body = {"success": success, "message": message}
     if ranks is not None:
